Force culture of the csv parser to french belgium
it appears that the culture of the file should be specified with the operations.
The issue was at tfs side where the culture was not the same as mine. it was then complaining about the format of some date times in the test 'ReadAFortisExportCsv'
waiting for a more global solution, I force that culture as the csv parsing.
Nevertheless it means that this is one important extra information that must be provided with a file to import.